+++ lib/features/settings/application/sidebar_cassette_spec/coordinators/settings_coordinator.dart
@@ -1,14 +1,8 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import '../../../../../essentials/sidebar/presentation/view_model/sidebar_cassette_card_view_model.dart';
-import '../../../../attachments/feature_level_providers.dart'
-    show
-        AttachmentArchiveAdoptionResult,
-        AttachmentArchiveAdoptionWorkflowState,
-        AttachmentArchiveLocationState;
 import '../../../domain/spec_classes/settings_cassette_spec.dart';
 import '../payloads/historical_archives_settings_cassette_payload.dart';
-import '../resolvers/attachment_archive_settings_resolver.dart';
 import '../resolvers/historical_archives_settings_resolver.dart';
 import '../resolvers/message_history_coverage_settings_resolver.dart';
 import '../resolvers/reset_message_data_settings_resolver.dart';
@@ -26,9 +20,6 @@ class SettingsCassetteCoordinator extends _$SettingsCassetteCoordinator {
     SettingsCassetteSpec spec, {
     required int cassetteIndex,
     List<HistoricalArchiveSidebarSourceSummary>? historicalArchivesKnownSources,
-    AttachmentArchiveLocationState? attachmentArchiveLocation,
-    AttachmentArchiveAdoptionWorkflowState? attachmentArchiveWorkflow,
-    AttachmentArchiveAdoptionResult? attachmentArchiveRecovery,
   }) async {
     return spec.when(
       historicalArchivesOverview: () => ref
@@ -58,14 +49,6 @@ class SettingsCassetteCoordinator extends _$SettingsCassetteCoordinator {
       imageSizeInfo: () => ref
           .read(settingsInfoResolverProvider.notifier)
           .resolve(title: 'Image Size', bodyText: 'Coming soon'),
-      attachmentArchive: () => ref
-          .read(attachmentArchiveSettingsResolverProvider.notifier)
-          .resolve(
-            cassetteIndex: cassetteIndex,
-            location: attachmentArchiveLocation!,
-            workflow: attachmentArchiveWorkflow!,
-            recovery: attachmentArchiveRecovery,
-          ),
     );
   }
 }