+++ lib/features/settings/application/view_spec/coordinators/view_spec_coordinator.dart
@@ -1,10 +1,7 @@
-import 'package:flutter/widgets.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import '../../../domain/spec_classes/settings_view_spec.dart';
-import '../resolvers/attachment_archive_panel_resolver.dart';
-import '../resolvers/historical_archives_panel_resolver.dart';
-import '../resolvers/message_history_coverage_report_panel_resolver.dart';
+import '../payloads/settings_panel_render_descriptor.dart';
 
 part 'view_spec_coordinator.g.dart';
 
@@ -15,14 +12,14 @@ class ViewSpecCoordinator extends _$ViewSpecCoordinator {
     // Stateless coordinator.
   }
 
-  Widget buildForSpec(SettingsViewSpec spec) {
+  SettingsPanelRenderDescriptor resolveForSpec(SettingsViewSpec spec) {
     return spec.when(
       attachmentArchiveWorkflow: () =>
-          AttachmentArchivePanelResolver().resolve(),
+          SettingsPanelRenderDescriptor.attachmentArchiveWorkflow,
       historicalArchivesWorkflow: () =>
-          HistoricalArchivesPanelResolver().resolve(),
+          SettingsPanelRenderDescriptor.historicalArchivesWorkflow,
       messageHistoryCoverageReport: () =>
-          MessageHistoryCoverageReportPanelResolver().resolve(),
+          SettingsPanelRenderDescriptor.messageHistoryCoverageReport,
     );
   }
 }