+++ lib/features/settings/application/view_spec/resolvers/attachment_archive_panel_resolver.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-import '../../../presentation/view/attachment_archive_panel.dart';
-
-class AttachmentArchivePanelResolver {
-  Widget resolve() {
-    return const AttachmentArchivePanel();
-  }
-}