+++ lib/features/settings/domain/spec_classes/settings_cassette_spec.freezed.dart
@@ -55,7 +55,7 @@ extension SettingsCassetteSpecPatterns on SettingsCassetteSpec {
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeMap<TResult extends Object?>({TResult Function( _HistoricalArchivesOverview value)?  historicalArchivesOverview,TResult Function( _MessageHistoryCoverageOverview value)?  messageHistoryCoverageOverview,TResult Function( _MessageHistoryCoverageHowToRead value)?  messageHistoryCoverageHowToRead,TResult Function( _MessageHistoryCoverageOlderMessagesNote value)?  messageHistoryCoverageOlderMessagesNote,TResult Function( _SendLogsPanel value)?  sendLogsPanel,TResult Function( _ResetMessageDataPanel value)?  resetMessageDataPanel,TResult Function( _TextSizeInfo value)?  textSizeInfo,TResult Function( _ImageSizeInfo value)?  imageSizeInfo,TResult Function( _AttachmentArchive value)?  attachmentArchive,required TResult orElse(),}){
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({TResult Function( _HistoricalArchivesOverview value)?  historicalArchivesOverview,TResult Function( _MessageHistoryCoverageOverview value)?  messageHistoryCoverageOverview,TResult Function( _MessageHistoryCoverageHowToRead value)?  messageHistoryCoverageHowToRead,TResult Function( _MessageHistoryCoverageOlderMessagesNote value)?  messageHistoryCoverageOlderMessagesNote,TResult Function( _SendLogsPanel value)?  sendLogsPanel,TResult Function( _ResetMessageDataPanel value)?  resetMessageDataPanel,TResult Function( _TextSizeInfo value)?  textSizeInfo,TResult Function( _ImageSizeInfo value)?  imageSizeInfo,required TResult orElse(),}){
 final _that = this;
 switch (_that) {
 case _HistoricalArchivesOverview() when historicalArchivesOverview != null:
@@ -66,8 +66,7 @@ return messageHistoryCoverageOlderMessagesNote(_that);case _SendLogsPanel() when
 return sendLogsPanel(_that);case _ResetMessageDataPanel() when resetMessageDataPanel != null:
 return resetMessageDataPanel(_that);case _TextSizeInfo() when textSizeInfo != null:
 return textSizeInfo(_that);case _ImageSizeInfo() when imageSizeInfo != null:
-return imageSizeInfo(_that);case _AttachmentArchive() when attachmentArchive != null:
-return attachmentArchive(_that);case _:
+return imageSizeInfo(_that);case _:
   return orElse();
 
 }
@@ -85,7 +84,7 @@ return attachmentArchive(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult map<TResult extends Object?>({required TResult Function( _HistoricalArchivesOverview value)  historicalArchivesOverview,required TResult Function( _MessageHistoryCoverageOverview value)  messageHistoryCoverageOverview,required TResult Function( _MessageHistoryCoverageHowToRead value)  messageHistoryCoverageHowToRead,required TResult Function( _MessageHistoryCoverageOlderMessagesNote value)  messageHistoryCoverageOlderMessagesNote,required TResult Function( _SendLogsPanel value)  sendLogsPanel,required TResult Function( _ResetMessageDataPanel value)  resetMessageDataPanel,required TResult Function( _TextSizeInfo value)  textSizeInfo,required TResult Function( _ImageSizeInfo value)  imageSizeInfo,required TResult Function( _AttachmentArchive value)  attachmentArchive,}){
+@optionalTypeArgs TResult map<TResult extends Object?>({required TResult Function( _HistoricalArchivesOverview value)  historicalArchivesOverview,required TResult Function( _MessageHistoryCoverageOverview value)  messageHistoryCoverageOverview,required TResult Function( _MessageHistoryCoverageHowToRead value)  messageHistoryCoverageHowToRead,required TResult Function( _MessageHistoryCoverageOlderMessagesNote value)  messageHistoryCoverageOlderMessagesNote,required TResult Function( _SendLogsPanel value)  sendLogsPanel,required TResult Function( _ResetMessageDataPanel value)  resetMessageDataPanel,required TResult Function( _TextSizeInfo value)  textSizeInfo,required TResult Function( _ImageSizeInfo value)  imageSizeInfo,}){
 final _that = this;
 switch (_that) {
 case _HistoricalArchivesOverview():
@@ -96,8 +95,7 @@ return messageHistoryCoverageOlderMessagesNote(_that);case _SendLogsPanel():
 return sendLogsPanel(_that);case _ResetMessageDataPanel():
 return resetMessageDataPanel(_that);case _TextSizeInfo():
 return textSizeInfo(_that);case _ImageSizeInfo():
-return imageSizeInfo(_that);case _AttachmentArchive():
-return attachmentArchive(_that);case _:
+return imageSizeInfo(_that);case _:
   throw StateError('Unexpected subclass');
 
 }
@@ -114,7 +112,7 @@ return attachmentArchive(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>({TResult? Function( _HistoricalArchivesOverview value)?  historicalArchivesOverview,TResult? Function( _MessageHistoryCoverageOverview value)?  messageHistoryCoverageOverview,TResult? Function( _MessageHistoryCoverageHowToRead value)?  messageHistoryCoverageHowToRead,TResult? Function( _MessageHistoryCoverageOlderMessagesNote value)?  messageHistoryCoverageOlderMessagesNote,TResult? Function( _SendLogsPanel value)?  sendLogsPanel,TResult? Function( _ResetMessageDataPanel value)?  resetMessageDataPanel,TResult? Function( _TextSizeInfo value)?  textSizeInfo,TResult? Function( _ImageSizeInfo value)?  imageSizeInfo,TResult? Function( _AttachmentArchive value)?  attachmentArchive,}){
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>({TResult? Function( _HistoricalArchivesOverview value)?  historicalArchivesOverview,TResult? Function( _MessageHistoryCoverageOverview value)?  messageHistoryCoverageOverview,TResult? Function( _MessageHistoryCoverageHowToRead value)?  messageHistoryCoverageHowToRead,TResult? Function( _MessageHistoryCoverageOlderMessagesNote value)?  messageHistoryCoverageOlderMessagesNote,TResult? Function( _SendLogsPanel value)?  sendLogsPanel,TResult? Function( _ResetMessageDataPanel value)?  resetMessageDataPanel,TResult? Function( _TextSizeInfo value)?  textSizeInfo,TResult? Function( _ImageSizeInfo value)?  imageSizeInfo,}){
 final _that = this;
 switch (_that) {
 case _HistoricalArchivesOverview() when historicalArchivesOverview != null:
@@ -125,8 +123,7 @@ return messageHistoryCoverageOlderMessagesNote(_that);case _SendLogsPanel() when
 return sendLogsPanel(_that);case _ResetMessageDataPanel() when resetMessageDataPanel != null:
 return resetMessageDataPanel(_that);case _TextSizeInfo() when textSizeInfo != null:
 return textSizeInfo(_that);case _ImageSizeInfo() when imageSizeInfo != null:
-return imageSizeInfo(_that);case _AttachmentArchive() when attachmentArchive != null:
-return attachmentArchive(_that);case _:
+return imageSizeInfo(_that);case _:
   return null;
 
 }
@@ -143,7 +140,7 @@ return attachmentArchive(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({TResult Function()?  historicalArchivesOverview,TResult Function()?  messageHistoryCoverageOverview,TResult Function()?  messageHistoryCoverageHowToRead,TResult Function()?  messageHistoryCoverageOlderMessagesNote,TResult Function()?  sendLogsPanel,TResult Function()?  resetMessageDataPanel,TResult Function()?  textSizeInfo,TResult Function()?  imageSizeInfo,TResult Function()?  attachmentArchive,required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({TResult Function()?  historicalArchivesOverview,TResult Function()?  messageHistoryCoverageOverview,TResult Function()?  messageHistoryCoverageHowToRead,TResult Function()?  messageHistoryCoverageOlderMessagesNote,TResult Function()?  sendLogsPanel,TResult Function()?  resetMessageDataPanel,TResult Function()?  textSizeInfo,TResult Function()?  imageSizeInfo,required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _HistoricalArchivesOverview() when historicalArchivesOverview != null:
 return historicalArchivesOverview();case _MessageHistoryCoverageOverview() when messageHistoryCoverageOverview != null:
@@ -153,8 +150,7 @@ return messageHistoryCoverageOlderMessagesNote();case _SendLogsPanel() when send
 return sendLogsPanel();case _ResetMessageDataPanel() when resetMessageDataPanel != null:
 return resetMessageDataPanel();case _TextSizeInfo() when textSizeInfo != null:
 return textSizeInfo();case _ImageSizeInfo() when imageSizeInfo != null:
-return imageSizeInfo();case _AttachmentArchive() when attachmentArchive != null:
-return attachmentArchive();case _:
+return imageSizeInfo();case _:
   return orElse();
 
 }
@@ -172,7 +168,7 @@ return attachmentArchive();case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>({required TResult Function()  historicalArchivesOverview,required TResult Function()  messageHistoryCoverageOverview,required TResult Function()  messageHistoryCoverageHowToRead,required TResult Function()  messageHistoryCoverageOlderMessagesNote,required TResult Function()  sendLogsPanel,required TResult Function()  resetMessageDataPanel,required TResult Function()  textSizeInfo,required TResult Function()  imageSizeInfo,required TResult Function()  attachmentArchive,}) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>({required TResult Function()  historicalArchivesOverview,required TResult Function()  messageHistoryCoverageOverview,required TResult Function()  messageHistoryCoverageHowToRead,required TResult Function()  messageHistoryCoverageOlderMessagesNote,required TResult Function()  sendLogsPanel,required TResult Function()  resetMessageDataPanel,required TResult Function()  textSizeInfo,required TResult Function()  imageSizeInfo,}) {final _that = this;
 switch (_that) {
 case _HistoricalArchivesOverview():
 return historicalArchivesOverview();case _MessageHistoryCoverageOverview():
@@ -182,8 +178,7 @@ return messageHistoryCoverageOlderMessagesNote();case _SendLogsPanel():
 return sendLogsPanel();case _ResetMessageDataPanel():
 return resetMessageDataPanel();case _TextSizeInfo():
 return textSizeInfo();case _ImageSizeInfo():
-return imageSizeInfo();case _AttachmentArchive():
-return attachmentArchive();case _:
+return imageSizeInfo();case _:
   throw StateError('Unexpected subclass');
 
 }
@@ -200,7 +195,7 @@ return attachmentArchive();case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>({TResult? Function()?  historicalArchivesOverview,TResult? Function()?  messageHistoryCoverageOverview,TResult? Function()?  messageHistoryCoverageHowToRead,TResult? Function()?  messageHistoryCoverageOlderMessagesNote,TResult? Function()?  sendLogsPanel,TResult? Function()?  resetMessageDataPanel,TResult? Function()?  textSizeInfo,TResult? Function()?  imageSizeInfo,TResult? Function()?  attachmentArchive,}) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>({TResult? Function()?  historicalArchivesOverview,TResult? Function()?  messageHistoryCoverageOverview,TResult? Function()?  messageHistoryCoverageHowToRead,TResult? Function()?  messageHistoryCoverageOlderMessagesNote,TResult? Function()?  sendLogsPanel,TResult? Function()?  resetMessageDataPanel,TResult? Function()?  textSizeInfo,TResult? Function()?  imageSizeInfo,}) {final _that = this;
 switch (_that) {
 case _HistoricalArchivesOverview() when historicalArchivesOverview != null:
 return historicalArchivesOverview();case _MessageHistoryCoverageOverview() when messageHistoryCoverageOverview != null:
@@ -210,8 +205,7 @@ return messageHistoryCoverageOlderMessagesNote();case _SendLogsPanel() when send
 return sendLogsPanel();case _ResetMessageDataPanel() when resetMessageDataPanel != null:
 return resetMessageDataPanel();case _TextSizeInfo() when textSizeInfo != null:
 return textSizeInfo();case _ImageSizeInfo() when imageSizeInfo != null:
-return imageSizeInfo();case _AttachmentArchive() when attachmentArchive != null:
-return attachmentArchive();case _:
+return imageSizeInfo();case _:
   return null;
 
 }
@@ -467,38 +461,6 @@ int get hashCode => runtimeType.hashCode;
 @override
 String toString() {
   return 'SettingsCassetteSpec.imageSizeInfo()';
-}
-
-
-}
-
-
-
-
-/// @nodoc
-
-
-class _AttachmentArchive implements SettingsCassetteSpec {
-  const _AttachmentArchive();
-  
-
-
-
-
-
-
-@override
-bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _AttachmentArchive);
-}
-
-
-@override
-int get hashCode => runtimeType.hashCode;
-
-@override
-String toString() {
-  return 'SettingsCassetteSpec.attachmentArchive()';
 }